refactor: consolidate and relocate scattered configuration utilities (Issue #282)
**Phase 1: Root-Level Files Consolidation**
- Moved p3.py to infra/p3/p3.py with delegation system
- Relocated pytest.ini to tests/pytest.ini
- Moved CHANGELOG.md to infra/docs/CHANGELOG.md
- Updated pixi.toml script paths

**Phase 2: L1/L2 Module Structure**
- Achieved clean root directory with essential files only
- Proper modular organization in infrastructure directories
- Maintained backward compatibility throughout

**Validation Completed**
- P3 CLI functionality preserved
- Import paths validated
- SSOT I/O enforcement followed
- Configuration centralization maintained
