+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-# DEPRECATED - File Removed
+# DEPRECATED - File Moved
 
-This file has been removed as part of GitHub issue #232 root directory cleanup.
+This file has been moved to infra/docs/CHANGELOG.md as part of GitHub issue #282 root directory cleanup.
 
-**Reason for removal**: Changelog content should be managed via GitHub releases and issue tracking per CLAUDE.md policy, not standalone .md files.
+**New location**: infra/docs/CHANGELOG.md
 
-**Content preserved**: Historical changelog information has been preserved in git history and should be managed through proper GitHub release processes going forward.
+**Reason for move**: Following L1/L2 module architecture - documentation files belong in infra/docs/ directory, not root.
 
-**Policy reference**: See CLAUDE.md - "Use GitHub Issues for ALL planning and documentation, NOT additional .md files"
+**Note**: As per CLAUDE.md policy, changelog content should be managed via GitHub releases and issue tracking, not standalone .md files.
+
+**Policy reference**: See CLAUDE.md - "MODULAR FILE PLACEMENT: All files must be in appropriate L1/L2 directories - avoid root directory unless essential"
