+++ infra/docs/CHANGELOG.md
@@ -1,0 +1,9 @@
+# DEPRECATED - File Removed
+
+This file has been removed as part of GitHub issue #232 root directory cleanup.
+
+**Reason for removal**: Changelog content should be managed via GitHub releases and issue tracking per CLAUDE.md policy, not standalone .md files.
+
+**Content preserved**: Historical changelog information has been preserved in git history and should be managed through proper GitHub release processes going forward.
+
+**Policy reference**: See CLAUDE.md - "Use GitHub Issues for ALL planning and documentation, NOT additional .md files"
