+++ infra/docs/README.md
@@ -1,0 +1,25 @@
+# Infrastructure Documentation
+
+Documentation files for infrastructure components and system administration.
+
+## Contents
+
+- **CHANGELOG.md** - Historical changelog (deprecated, see GitHub releases)
+
+## Purpose
+
+This directory contains infrastructure-related documentation that has been moved from the root directory as part of Issue #282 modular file placement policy.
+
+**Policy Compliance**: Per CLAUDE.md - "MODULAR FILE PLACEMENT: All files must be in appropriate L1/L2 directories - avoid root directory unless essential"
+
+## Documentation Standards
+
+- **Primary Documentation**: Use GitHub Issues for planning and tracking
+- **Technical Documentation**: Module-specific README files in respective directories
+- **Historical Documents**: Preserved here for reference but deprecated in favor of GitHub releases
+
+## Related Documentation
+
+- **Main Documentation**: [README.md](../../README.md) - Project overview and architecture
+- **Governance**: [CLAUDE.md](../../CLAUDE.md) - Company policies and procedures
+- **Infrastructure**: [infra/README.md](../README.md) - Infrastructure module documentation
