+++ infra/p3/__init__.py
@@ -6,3 +6,7 @@ P3 CLI maintenance, development, and optimization tools.
 Per CLAUDE.md organizational authority, infra-ops-agent has exclusive technical
 responsibility for P3 CLI codebase maintenance and enhancement.
 """
+
+from .p3 import P3CLI, get_version_string, main
+
+__all__ = ["P3CLI", "get_version_string", "main"]
