+++ p3.py
@@ -1,191 +1,22 @@
 #!/usr/bin/env python3
 """
-P3 CLI - Simplified Workflow-Oriented Command System
-Only 9 essential workflow commands for developer productivity
+P3 CLI - Root Entry Point
+Delegates to the actual P3 implementation in infra/p3/
 """
 
-import os
-import subprocess
 import sys
-import time
 from pathlib import Path
-from typing import Dict, Optional
 
+# Add the project root to Python path for import
+project_root = Path(__file__).parent
+sys.path.insert(0, str(project_root))
 
-def get_version_string():
-    return "2.0.1-simplified"
+try:
+    from infra.p3.p3 import main
 
-
-class P3CLI:
-    """Simplified P3 CLI with only 9 workflow commands."""
-
-    def __init__(self):
-        self.project_root = Path.cwd()
-        self.commands = self._load_commands()
-
-    def _get_current_branch(self) -> str:
-        """Get current git branch."""
-        try:
-            result = subprocess.run(
-                ["git", "branch", "--show-current"], capture_output=True, text=True, timeout=5
-            )
-            return result.stdout.strip() if result.returncode == 0 else "unknown"
-        except:
-            return "unknown"
-
-    def _load_commands(self) -> Dict[str, str]:
-        """Load the essential workflow commands."""
-        return {
-            # Core Workflow Commands (8 total)
-            "ready": "python infra/system/workflow_ready.py",  # Start working
-            "stop": "python infra/system/workflow_stop.py",  # Stop working (release resources)
-            "reset": "python infra/system/workflow_reset.py",  # Fix environment
-            "check": "python infra/development/workflow_check.py",  # Validate code
-            "test": "python infra/scripts/utilities/run_test.py",  # Test
-            "ship": "python infra/workflows/pr_creation.py",  # Create PR
-            "build": "python ETL/build_dataset.py",  # Build dataset
-            "version": "version_command",  # Version info
-        }
-
-    def run(self, command: str, args: list):
-        """Execute a P3 command."""
-        start_time = time.time()
-        print(f"🚀 P3 CLI v{get_version_string()} - Command: {command}")
-        print(f"⏰ Start time: {time.strftime('%Y-%m-%d %H:%M:%S')}")
-
-        if command == "help" or command is None:
-            self.print_help()
-            return
-
-        if command not in self.commands:
-            print(f"❌ Unknown command: {command}")
-            print("Available commands: ready, stop, reset, check, test, ship, build, version")
-            print("Use 'p3 help' for details")
-            sys.exit(1)
-
-        # Handle version command
-        if command == "version":
-            print(f"P3 Version: {get_version_string()}")
-            print(f"   Script: {Path(__file__).resolve()}")
-            print(f"   Working directory: {self.project_root}")
-            print(f"   Git branch: {self._get_current_branch()}")
-            return
-
-        # Handle ship command with special arguments
-        if command == "ship":
-            if len(args) < 2:
-                print("❌ Error: title and issue number required")
-                print('Usage: p3 ship "PR Title" ISSUE_NUMBER')
-                sys.exit(1)
-            cmd_string = f'pixi run {self.commands[command]} "{args[0]}" {args[1]}'
-        elif command in ["build", "test", "check"]:
-            scope = args[0] if args else "f2"
-            cmd_string = f"pixi run {self.commands[command]} {scope}"
-            print(f"📋 Using scope: {scope}")
-        else:
-            cmd_string = f"pixi run {self.commands[command]}"
-            if args:
-                cmd_string += " " + " ".join(args)
-
-        # Execute command with detailed logging
-        print(f"🔧 Command: {self.commands[command]}")
-        print(f"🚀 Executing: {cmd_string}")
-        print(f"📁 Working directory: {self.project_root}")
-        print(f"🌿 Git branch: {self._get_current_branch()}")
-        print("=" * 50)
-
-        os.chdir(self.project_root)
-
-        # Log execution start
-        exec_start = time.time()
-        print(f"⚡ Command execution started at {time.strftime('%H:%M:%S')}")
-
-        result = subprocess.run(cmd_string, shell=True)
-
-        # Log execution completion
-        exec_end = time.time()
-        exec_duration = exec_end - exec_start
-        total_duration = exec_end - start_time
-
-        print("=" * 50)
-        print(f"⏱️  Command execution time: {exec_duration:.2f}s")
-        print(f"⏱️  Total P3 time: {total_duration:.2f}s")
-        print(f"🏁 Finished at: {time.strftime('%Y-%m-%d %H:%M:%S')}")
-        print(f"🔄 Exit code: {result.returncode}")
-
-        # Provide helpful diagnostics for failures
-        if result.returncode != 0:
-            print()
-            print("🔍 COMMAND FAILED - Quick Diagnostics:")
-            print(f"   Command: {command}")
-            print(f"   Exit code: {result.returncode}")
-            print()
-            print("💡 Common solutions:")
-            if command == "ready":
-                print("   • Check if Podman machine is running: podman machine list")
-                print("   • Try: p3 reset (clean restart)")
-                print("   • Check system resources: Activity Monitor")
-            elif command == "test":
-                print("   • Check if environment is ready: p3 ready")
-                print("   • Try smaller scope: p3 test f2")
-                print("   • Check dependencies: pixi install")
-            elif command == "ship":
-                print("   • Ensure tests pass first: p3 test f2")
-                print("   • Check git status and commit changes")
-                print("   • Verify issue number is correct")
-            elif command == "check":
-                print("   • Check code formatting issues")
-                print("   • Try fixing with: p3 ready")
-                print("   • Check pixi environment: pixi install")
-            else:
-                print("   • Try: p3 ready (setup environment)")
-                print("   • Check logs above for specific errors")
-                print("   • Use 'p3 reset' for major issues")
-
-        sys.exit(result.returncode)
-
-    def print_help(self):
-        """Print help message."""
-        print(
-            f"""
-🚀 P3 CLI - Workflow-Oriented Development Commands
-
-DAILY WORKFLOW (5 commands):
-  p3 ready                  Start working (env + services)
-  p3 stop [--full] [--force] Stop working (release resources)
-  p3 check [scope]          Validate code (format + lint + basic tests)
-  p3 test [scope]           Unit tests + integration + e2e (superset of CI)
-  p3 ship "title" issue     Publish work (comprehensive test + PR + cleanup)
-
-TROUBLESHOOTING (1 command):
-  p3 reset                  Fix environment (clean restart)
-
-DATA & VERSION (2 commands):
-  p3 build [scope]          Build dataset (f2/m7/n100/v3k)
-  p3 version                Show version
-
-STOP OPTIONS:
-  --full                    Stop Podman machine (complete shutdown)
-  --force                   Force stop without graceful shutdown
-
-SCOPES:
-  f2    Fast 2 companies (development)
-  m7    Magnificent 7 (testing)
-  n100  NASDAQ 100 (validation)  
-  v3k   Russell 3000 (production)
-
-Version: {get_version_string()}
-"""
-        )
-
-
-def main():
-    """Main entry point."""
-    cli = P3CLI()
-    command = sys.argv[1] if len(sys.argv) > 1 else None
-    args = sys.argv[2:] if len(sys.argv) > 2 else []
-    cli.run(command, args)
-
-
-if __name__ == "__main__":
-    main()
+    if __name__ == "__main__":
+        main()
+except ImportError as e:
+    print(f"❌ Error importing P3 CLI: {e}")
+    print("🔧 Please ensure the infra/p3/ module is properly installed")
+    sys.exit(1)
