+++ pytest.ini
@@ -1,54 +1,9 @@
-[tool:pytest]
-# Pytest configuration for consistent testing across environments
+# DEPRECATED - File Moved
 
-# Test discovery patterns
-python_files = test_*.py *_test.py
-python_classes = Test*
-python_functions = test_*
+This file has been moved to tests/pytest.ini as part of GitHub issue #282 root directory cleanup.
 
-# Test directories - modular testing architecture (Issue #282)
-testpaths =
-    tests
-    common/tests
-    ETL/tests
-    engine/tests
-    evaluation/tests
-    infra/tests
+**New location**: tests/pytest.ini
 
-# Minimum version
-minversion = 8.0
+**Reason for move**: Following L1/L2 module architecture - testing configuration belongs in tests/ directory, not root.
 
-# Add current directory to Python path and configure options
-addopts = 
-    --tb=short
-    --strict-markers
-    --strict-config
-    --verbose
-    --maxfail=10
-    --disable-warnings
-    --cov-report=term-missing 
-    --cov-fail-under=40
-
-# Markers for test categorization (CI-aligned for p3 test superset)
-markers =
-    unit: Unit tests for individual components (CI priority) - co-located with modules
-    integration: Integration tests for cross-module functionality (CI secondary) - root tests/ directory
-    slow: Slow tests that may take longer to run
-    core: Tests for core system components (CI critical)
-    agents: Tests for agent-related modules
-    build: Tests for build and quality modules (CI critical)
-    monitoring: Tests for monitoring components
-    schemas: Tests for schema definitions (CI critical)
-    utils: Tests for utility modules
-    legacy: Tests for legacy/deprecated components
-    fast: Fast tests for development (p3 check scope)
-    ci: Tests that exactly match CI requirements (p3 test scope)
-
-# Test timeout (prevent hanging tests)
-timeout = 300
-
-# Filtering options
-filterwarnings = 
-    ignore::UserWarning
-    ignore::DeprecationWarning
-    ignore::PendingDeprecationWarning
+**Policy reference**: See CLAUDE.md - "MODULAR FILE PLACEMENT: All files must be in appropriate L1/L2 directories - avoid root directory unless essential"
